shared.types module renamed
