+++ apps/data_handler/handlers/loan_states/nostra_alpha/events.py
@@ -26,8 +26,8 @@ from shared.helpers import add_leading_zeros
 from shared.loan_entity import LoanEntity
 from shared.starknet_client import StarknetClient
 from shared.state import State
-from shared.types import InterestRateModels, Portfolio, Prices, TokenParameters
-from shared.types.nostra import (
+from shared.custom_types import InterestRateModels, Portfolio, Prices, TokenParameters
+from shared.custom_types.nostra import (
     NostraAlphaCollateralTokenParameters,
     NostraDebtTokenParameters,
 )
