+++ apps/shared/amms.py
@@ -7,7 +7,7 @@ from typing import Dict, List, Optional
 from shared.blockchain_call import balance_of, func_call, get_myswap_pool
 from shared.constants import POOL_MAPPING, TOKEN_SETTINGS
 from shared.helpers import add_leading_zeros
-from shared.types import TokenSettings, TokenValues
+from shared.custom_types import TokenSettings, TokenValues
 
 
 class Pair:
