+++ apps/shared/constants.py
@@ -4,7 +4,7 @@ from decimal import Decimal
 from enum import Enum
 from typing import List, Union
 
-from shared.types import TokenSettings
+from shared.custom_types import TokenSettings
 
 ZKLEND = "zkLend"
 NOSTRA_ALPHA = "Nostra Alpha"
