+++ apps/shared/custom_types/zklend.py
@@ -1,7 +1,7 @@
 from collections import defaultdict
 from dataclasses import dataclass
 
-from shared.types import BaseTokenParameters
+from shared.custom_types import BaseTokenParameters
 
 
 class ZkLendCollateralEnabled(defaultdict):
